requirements/requirements.txt
aniso8601==8.0.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.4
pbr==5.4.3
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
stevedore==1.31.0
urllib3==1.25.3
Werkzeug==0.16.0